+++ src/hooks/mutations/useRevokeIssuedCouponMutation.ts
@@ -1,18 +1,8 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { toast } from "react-toastify";
+import { useMutation } from "@tanstack/react-query";
 import { couponApi } from "@/apis/couponApi";
-import { QueryKey } from "@/constants/queryKey";
 
 export default function useRevokeIssuedCouponMutation() {
-  const queryClient = useQueryClient();
-
   return useMutation({
     mutationFn: couponApi.revokeIssuedCoupon,
-    onSuccess: () => {
-      queryClient.invalidateQueries({
-        queryKey: [QueryKey.issuedCouponList],
-      });
-      toast.success("쿠폰을 회수하였습니다.");
-    },
   });
 }
